Generate the CRDs using the JSON schema to prevent number -> int type changes for TF AWS provider v4.67.0.

+++ apis/apigatewayv2/v1beta1/zz_stage_types.go
@@ -60,7 +60,7 @@ type DefaultRouteSettingsInitParameters struct {
 	LoggingLevel *string `json:"loggingLevel,omitempty" tf:"logging_level,omitempty"`
 
 	// Throttling burst limit for the default route.
-	ThrottlingBurstLimit *int64 `json:"throttlingBurstLimit,omitempty" tf:"throttling_burst_limit,omitempty"`
+	ThrottlingBurstLimit *float64 `json:"throttlingBurstLimit,omitempty" tf:"throttling_burst_limit,omitempty"`
 
 	// Throttling rate limit for the default route.
 	ThrottlingRateLimit *float64 `json:"throttlingRateLimit,omitempty" tf:"throttling_rate_limit,omitempty"`
@@ -80,7 +80,7 @@ type DefaultRouteSettingsObservation struct {
 	LoggingLevel *string `json:"loggingLevel,omitempty" tf:"logging_level,omitempty"`
 
 	// Throttling burst limit for the default route.
-	ThrottlingBurstLimit *int64 `json:"throttlingBurstLimit,omitempty" tf:"throttling_burst_limit,omitempty"`
+	ThrottlingBurstLimit *float64 `json:"throttlingBurstLimit,omitempty" tf:"throttling_burst_limit,omitempty"`
 
 	// Throttling rate limit for the default route.
 	ThrottlingRateLimit *float64 `json:"throttlingRateLimit,omitempty" tf:"throttling_rate_limit,omitempty"`
@@ -104,7 +104,7 @@ type DefaultRouteSettingsParameters struct {
 
 	// Throttling burst limit for the default route.
 	// +kubebuilder:validation:Optional
-	ThrottlingBurstLimit *int64 `json:"throttlingBurstLimit,omitempty" tf:"throttling_burst_limit,omitempty"`
+	ThrottlingBurstLimit *float64 `json:"throttlingBurstLimit,omitempty" tf:"throttling_burst_limit,omitempty"`
 
 	// Throttling rate limit for the default route.
 	// +kubebuilder:validation:Optional
@@ -128,7 +128,7 @@ type RouteSettingsInitParameters struct {
 	RouteKey *string `json:"routeKey,omitempty" tf:"route_key,omitempty"`
 
 	// Throttling burst limit for the route.
-	ThrottlingBurstLimit *int64 `json:"throttlingBurstLimit,omitempty" tf:"throttling_burst_limit,omitempty"`
+	ThrottlingBurstLimit *float64 `json:"throttlingBurstLimit,omitempty" tf:"throttling_burst_limit,omitempty"`
 
 	// Throttling rate limit for the route.
 	ThrottlingRateLimit *float64 `json:"throttlingRateLimit,omitempty" tf:"throttling_rate_limit,omitempty"`
@@ -151,7 +151,7 @@ type RouteSettingsObservation struct {
 	RouteKey *string `json:"routeKey,omitempty" tf:"route_key,omitempty"`
 
 	// Throttling burst limit for the route.
-	ThrottlingBurstLimit *int64 `json:"throttlingBurstLimit,omitempty" tf:"throttling_burst_limit,omitempty"`
+	ThrottlingBurstLimit *float64 `json:"throttlingBurstLimit,omitempty" tf:"throttling_burst_limit,omitempty"`
 
 	// Throttling rate limit for the route.
 	ThrottlingRateLimit *float64 `json:"throttlingRateLimit,omitempty" tf:"throttling_rate_limit,omitempty"`
@@ -179,7 +179,7 @@ type RouteSettingsParameters struct {
 
 	// Throttling burst limit for the route.
 	// +kubebuilder:validation:Optional
-	ThrottlingBurstLimit *int64 `json:"throttlingBurstLimit,omitempty" tf:"throttling_burst_limit,omitempty"`
+	ThrottlingBurstLimit *float64 `json:"throttlingBurstLimit,omitempty" tf:"throttling_burst_limit,omitempty"`
 
 	// Throttling rate limit for the route.
 	// +kubebuilder:validation:Optional
